backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
python-dotenv>=0.19.0
groq>=0.37.1
speechrecognition>=3.8.1
pyttsx3>=2.90
python-multipart
httpx
